FIX : 역상성 값 0 출력되는 점 수정

+++ Assets/Scripts/Status/StatusActor.cs
@@ -87,7 +87,7 @@ public class StatusActor
             _statusElementList[i].StatusData.Calculate<T>(ref rate, ref value, _statusElementList[i].OverlapCount);
         }
         //Debug.Log(value + " " + rate);
-        return (int)(((float)value) * rate);
+        return value * rate;
     }
 
     public void AddStatusData(ICaster caster, StatusData statusData)
